DRY refactoring for the course selection form
During the lifecycle of the evaluation, a student representative can
at several points move multiple courses forward to the next stage
(prepared, approved, published).

This happens to improve the usability of the web front-end and to
reduce the number of emails sent to the bare minimum.

The forms used for this en-bloque course management are nearly
identical and are DRY refactored by this commit in order to have a
single code base.