Ensure relogin possibility
At the moment the application redirects the user with a "301 moved permanently" status code to the main site after a successful login. The 301 indicates that the browser should cache this information. If you need to reauthorize (after logout or inactivity), the browser may not do the request to /login and opens / directly.

Problem occurred: with chrome 27.0.1453.116 m on windows 8