Fix crash in user create form. Fixes #868, fixes #832
The problem was that dd42a3cf895fefb8d4b2abc1c46ea8eb825446ae changed an
`object` variable in the template to `user`. `user`, however, is by default set
to the logged in user. thus, when the form was created without instance (when
creating a new user), the logged in user was taken instead.

A resulting template crash did not crash in production since DEBUG=False
eats such errors.