Fix csrf errors on language change, fixes #862
The caching of the menu bar cached per user name. As a result, all anonymous users got the same menu bar, and a user got the same menu bar each time when they logged in. Now, the csrf token was part of the language change form in the menu bar and thus got cached, but should be different for each anonymous user and change on every login.