Save the language in the user's session on first visit
This resolves the following inconsistency: If a user first visits the page, the brower's language gets written to user.language, but is not saved in the session. If the user now changes the browser's language, EvaP's language will also change. However, if the user logs out and logs in again, changing the browser's language will not change EvaP's language, because on the second login, the user.language will be written to the session.

@He3lixxx tripped over this while writing tests for the results cache, see #1229.

In addition, this moves the LocaleMiddleware before the AuthenticationMiddleware, since set_or_get_language (called during login) used the active language, which is set by the AuthenticationMiddleware. Without this change, it was necessary to reset the language in the tests manually. A user-facing bug might occur if a user calls the login url directly, then they would probably have the language set that was currently active in the django thread.