Remove default ordering by name_de. Fixes #1243 (#1438)
For Semester and Questionnaire, i replaced it with pk. They are primarily ordered by other attributes and i figured if those are the same, any stable sorting would be enough.
For Course, I went through the places where the ordering is user-visible and added a manual sort() in the one place where I saw it was missing. In the EvaluationForm and SingleResultForm the course field is now completely unordered, but only when creating a new Evaluation/SingleResult, and I thought that would not be worth it.