Rewrite staff/importers.py to be more modular
Idea / Model is that importers start with the binary excel content and
then put their current state through a series of steps that each gets
them a bit closer to their result. On errors, it simply raises an error
aborting the import. After each step, checkers can be run to add errors
or warnings that are more user friendly. This should allow to understand
the existing checks and add new checks very easily.