Simplify let_user_vote_for_evaluation
This speeds up tests by about 10%.

The actual student vote view is tested in student/tests/test_views.py,
so it doesn't make sense to have ~20 tests go through that view. This
was especially bad since the vote view is not optimized, so for some
tests, this took more than a second.

Since we don't need the rendered pages, and most of the time, we don't
even need any answers, this new version should also do the trick, and is
about 100x to 1000x faster (10% improvement over all tests).