Use forms instead of ajax for textanswer review
The forms submit to the same endpoint which now returns http 204 on
success instead of 200. This instructs the browser not to reload the
page, which is equivalent to what it did before, but with fewer custom
behavior.

In the "full" view, there are now four buttons instead of three.
They correspond to the four states that a textanswer can be in.
Before, the fourth state "undecided" / "unreviewed" was displayed by all
buttons being inactive.

Additionally, I renamed the action "hide" to "delete" and the request
parameter "id" to "answer_id".