Remove `django-sendfile2` and use the builtin `FileResponse` instead
We don't even remember why we used `sendfile` in the first place, it
claims to be optimized for sending large files and has specialized
backends, but we only send smallish grade documents and use the "simple"
backend anyways. Content type detection works with `FileResponse` too,
so there really seems to be no reason to have the dependency anymore.