Add sourcemap for tom-select bootstrap css file
With Django 4.1, `collectstatic` apparently cares about sourcemaps and
complains about a missing one after the update. The css file for
tom-select's bootstrap css had the indicator comment for a corresponding
source map, but we didn't have the file. Removing the indicator comment
works too, but that would mean that we distribute a modififed version of
the css file.