Remove some redundant id parameters
It is quite common that we have views that take an id for an evaluation
and a semester such that the evaluation must belong to the semester.
I believe that such requirements should be avoided, because we need to
check that they hold everywhere anyways (or fail to do so). The only
nice thing about this is that users can the /evaluation/foo suffix from
the url to get to the corresponding semester - but in a normal
interaction they can already do that using the breadcrumbs.

I noticed that we are actually missing on such check for the
textanswer-update view (and abused that ourselves in the tests). An evil
reviewer could change the review decision for a textanswer after the
evaluation is published (or similar). Consequently, I removed all other
parameters from urls containing textanswer ids.