Remove functionally dependent ids from URLs
For example, we had something like `/semester/42/course/1337` before,
but the `/course/1337` already implies that the semester should be - the
course's semester. The motivation for this change is a (only staff) view
I found in another PR that checked some condition for a course _on the
semester_, but never ensured that the semester from the URL matched the
course's semester.

I reverted the change for `evaluation_operation`, because it is more
tricky there: There could be no evaluation sent so we wouldn't know
where to redirect the client to if the semester wasn't there anymore. It
would be great if we could just respond with some HTTP code that says
"you made a mistake, please reload your current URL", but I couldn't
find anything (`HTTP 205: Reset Content` only clears forms).