Provide development environment via `nix` (#2274)
I am only keeping the meaningful commit messages:

* Add bash completion in development shells

It turns out that nix shells have a nice way to load completions
already. This even uses "bash" in the installation, so that it does not
interfere with other shells that may load the nix environment with
direnv.

    https://www.github.com/NixOS/nixpkgs/pull/103501
    https://www.github.com/scop/bash-completion/blob/7f1165018a72f837d59b3cea7071efa0108841a2/README.md

Since I now used the "lazy" approach for loading the completions, the
completion on `python manage.py <Tab>` does not work anymore, because
the builtin completions are found first. We could overwrite
BASH_COMPLETION_USER_DIR to have the highest precedence, but I think
that most users use `./manage.py` anyways (it is what we recommend in
our README) and the current approach seems to be the most canonical way
to do this within nix packages.

* Use unix domain sockets instead of TCP for postgres and redis

* Don't use `projectDir`

This should avoid that editing some views.py file changes the hash of
the poetry environment and thus save on some rebuilds. (The rebuilds
would only create some symlinks, all actually expensive steps are
cached, but this change won't hurt).

* Remove `poetry` from packages

While `poetry2nix` adds an overlay on the Python package set that
disables running the tests, this was not true for the additional poetry
from nixpkgs directly. With the override, this poetry had to be rebuilt
and thus the tests ran again.

* Add gettext to npm-ci pc process too

This should fix msgfmt errors for `./manage.py compilemessages`.