add option to disable standard tomselect sorting
according to tomselect docs by default the result is sorted by $score so the value or group which matches the sorting best.
But we want to order the groups by our sorting with summer and winter terms.
The sortField option here disables that and maintains the original order of options
(see https://tom-select.js.org/docs/)