Refactor `nix/services.nix` so that `services` does not depend on venv
Even though `nix run .#services` disables the `init-django` job, it
still includes it in the job list, and thus needs to build all the nix
derivations related to it. With this change, this is changed so that the
jobs are only available in `services-full`. This should speed up the
backup-process CI job, because it does not need to build the Python
environment anymore.

You can check this by comparing the output of

    nix run nixpkgs#nix-tree -- --derivation .#packages.x86_64-linux.services
    nix run nixpkgs#nix-tree -- --derivation .#packages.x86_64-linux.services-full

for example by searching for "django".