Handle boolean inputs for `setup_evap` correctly
As it turns out, the inputs are strings and even when using `false`, the
steps "Install Node dependencies" and "Start database" would run.

After fixing this, it turns out some of our jobs didn't specify their
parameters correctly, and pylint failed because it couldn't open the log
file, which lives in the data/ directory.