log m2m change directly in user form submit
since its not possible to make a generic m2m reverse logging, we call the log_m2m_change on the correct evaluation from the submit.
the added and removed evaluations a determined by calculating the diff to the previous evaluations of the user.
field_name is currently not determined dynamically.