Publish EvaP to PyPI (#2328)
Closes #2303 

Since we don't want to use nix in production (for example because security updates might take a long time to propagate through nixpkgs / through our github release cycle), `pip install .` or `pip install evap` seems to be the most canonical way. This allows us to install security patches via `pip install --upgrade evap`.

If we really need to override a dependency version beyond the allowed range in `pyproject.toml`, we can install EvaP from sources outside of PyPI. The script to build the release artifacts is available as `nix run .#build-dist`. It is possible to run this on a development machine and copy the `.whl` file over to the server.

The release wheel includes compiled TypeScript, SCSS and translation files, so there is no need to install NodeJS or gettext on the production server.

The new release workflow is:
1. Add translations and commit them.
2. Bump version in `pyproject.toml`, run `nix develop .#impure --command uv lock`, commit it, tag it, push it.
3. After the CI is done on main, dispatch a release workflow run to publish the wheel to PyPI.
4. Create Github Release.

The versioning scheme we agreed on is year-month-patch, for example "2024.11.0". I suppose we should take care that we don't necessarily push breaking changes like updating the Python version in a new patch. Following the Python Packaging User Guide, we don't include leading zeros in release versions like "2025.4.0".

Since upgrading to the next stable revision of EvaP is handled by pip, there is no need for the `release` branch anymore.

See
- https://packaging.python.org/en/latest/tutorials/packaging-projects/
- https://docs.pypi.org/trusted-publishers/using-a-publisher/
- https://packaging.python.org/en/latest/specifications/version-specifiers/