Take `INSTITUTION_EMAIL_REPLACEMENTS` into account when checking email transitions
The current version wrongly reports that a user's email has changed,
even though we changed it ourselves in `UserProfile.save`. This is
because the user is queried with the email address cleaned with
`clean_email`, and saved with the cleaned email, but we compare the
stored email with the uncleaned one.

I now changed it so that the email is cleaned right after we fetch the
claims from the OIDC provider. This ensures that we only deal with
cleaned emails everywhere else.