Fix flaky `StudentVoteLiveTest.test_resolving_submit_errors_clears_warning` (#2492)
We noticed test failures in CI:

```plain
selenium.common.exceptions.StaleElementReferenceException: Message: The element with the reference b70abd94-a473-4dad-ad6f-a8629a9007ae is stale; either its node document is not the active document, or it is no longer connected to the DOM; For documentation on this error, please visit: https://www.selenium.dev/documentation/webdriver/troubleshooting/errors#staleelementreferenceexception
```

on the line

```python
checkbox = row.find_element(By.CSS_SELECTOR, "input[type=radio][value='2'] + label.choice-error")
```

From the documentation, it seems that the `row` variable becomes stale. This can occur if the browser is slow and the page reload only happens after we assigned the `row` variable initially. To fix this, I inserted a wait command that ensures that the page has reloaded. The other tests in this file don't seem to submit the form, so they should be fine.

Note that the CI failure can be reproduced by moving the `row = ` line before the `submit_button.click()` call.